+++ bin/main.rb
@@ -1,0 +1,3 @@
+require_relative '../lib/bot'
+
+NewsBot.new
